backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
transformers==4.40.1
torch==2.3.0
pydantic==2.7.1
google-genai
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0
